+++ nengo/utils/tests/test_logging.py
@@ -7,11 +7,11 @@ import nengo.utils.logging
 
 
 def test_log_to_console():
-    nengo.log(debug=False, path=None)
+    nengo.log(path=None)
     assert logging.root.getEffectiveLevel() == logging.WARNING
     assert nengo.utils.logging.console_handler in logging.root.handlers
     n_handlers = len(logging.root.handlers)
-    nengo.log(debug=True, path=None)
+    nengo.log('debug', path=None)
     assert logging.root.getEffectiveLevel() == logging.DEBUG
     assert len(logging.root.handlers) == n_handlers
     logging.root.handlers.remove(nengo.utils.logging.console_handler)
@@ -19,13 +19,13 @@ def test_log_to_console():
 
 def test_log_to_file(tmpdir):
     tmpfile = str(tmpdir.join("log.txt"))
-    nengo.log(debug=False, path=tmpfile)
+    nengo.log(path=tmpfile)
     n_handlers = len(logging.root.handlers)
     handler = logging.root.handlers[-1]
     assert logging.root.getEffectiveLevel() == logging.WARNING
     assert isinstance(handler, logging.FileHandler)
     assert handler.baseFilename == tmpfile
-    nengo.log(debug=True, path=tmpfile)
+    nengo.log('debug', path=tmpfile)
     assert logging.root.getEffectiveLevel() == logging.DEBUG
     assert len(logging.root.handlers) == n_handlers
     logging.root.handlers.remove(handler)